+++ Kernel/System/ProcessManagement/TransitionAction/TicketCustomerSet.pm
@@ -103,6 +103,7 @@ sub Run {
 
     # use ticket attributes if needed
     $Self->_ReplaceTicketAttributes(%Param);
+    $Self->_ReplaceAdditionalAttributes(%Param);
 
     if (
         !$Param{Config}->{CustomerID}