+++ Kernel/System/ProcessManagement/TransitionAction/TicketStateSet.pm
@@ -102,6 +102,7 @@ sub Run {
 
     # use ticket attributes if needed
     $Self->_ReplaceTicketAttributes(%Param);
+    $Self->_ReplaceAdditionalAttributes(%Param);
 
     if ( !$Param{Config}->{StateID} && !$Param{Config}->{State} ) {
         $Kernel::OM->Get('Kernel::System::Log')->Log(