+++ Kernel/System/ProcessManagement/TransitionAction/TicketTitleSet.pm
@@ -94,6 +94,7 @@ sub Run {
 
     # use ticket attributes if needed
     $Self->_ReplaceTicketAttributes(%Param);
+    $Self->_ReplaceAdditionalAttributes(%Param);
 
     # Check for required parameters in ConfigHash
     if ( !defined $Param{Config}->{Title} ) {