+++ Kernel/System/ProcessManagement/TransitionAction/TicketTypeSet.pm
@@ -97,6 +97,7 @@ sub Run {
 
     # use ticket attributes if needed
     $Self->_ReplaceTicketAttributes(%Param);
+    $Self->_ReplaceAdditionalAttributes(%Param);
 
     if ( !$Param{Config}->{TypeID} && !$Param{Config}->{Type} ) {
         $Kernel::OM->Get('Kernel::System::Log')->Log(